WHO WE ARE Our purpose  booksnsports SA was officially launched as an NPO in South africa in 2011 as a collaborative effort among young members of a sport group in port elizabeth. The initiative was led by our founder Patient Kalume and co-founder Daniel Taokubong with the aim of diversifying the group's activities to create meaningful and fun activities for youths in the area including but not restricted to sport. With that a robust organization that included sport, art, academic follow up like tutoring and media was created. the founder, being a tutor and a fitness trainer decided to officially transform the mouvement into an NPO after noticing the need of an adequate education amongst the youth.  booksnsports SA strives to stand out with a different approach to development work and vouch for democracy and transparency. Experience has taught us that for a project to have long-term effects, it is crucial to have the beneficiaries partake in every aspect of the activities.  We come from Port elizabeth.   Being that most of our members have grown up in Port elizabeth, we know first-hand the feeling of boredom and the feeling of having nothing meaningful to do. Idleness often is an enemy as threatening as poverty itself. Notwithstanding, port elizabeth is a vibrant, creative and lively area with strong and friendly communities. Rather the problem, especially for young people, is the lacking of tools, space and appropriate guidance to develop their creative interests. This reflects a general need for skill-developing and confidence-building, concerning education on other levels than merely the academic.  We engage the youth.  Although sport was the major foundation of our own personal development, we felt it was not by itself sufficient. The activities have therefore evolved into art,tutoring(youth that are still in schools) in area like mathematic and science , sport and an education program to pay school fees, provide mentoring sessions and individual academic follow up. These activities are good ways to sidestep destructive activities like criminality and substance abuse and also, it enhances the opportunities for youths to realize their potential and increase their chances and ability to build a great future for themselves.  It may be challenging but time has proven that it is possible.  booksnsports SA members are not exempted from the challenging situations that come from growing up in a slum. Over the years since it started as a soccer group and extending to where it is, a fully operating NPO, we have had the pleasure of witnessing some of our members growing from young, clueless but talented kids to self sufficient inspired adults. This image constantly reminds us to focus on participation and ownership as fundamentals of community-driven development. booksnsports is built upon the involvement of the youths, playing important roles both in everyday work and projects as well as in the overall managing of the organization.  Our Reach  Our organization continues to expand and develop. Our members meet at the greenfield ground for sports sessions every afternoon, six days a week and we have a daily attendance of about 30 youths . we gather as well with Members(about 20) that needs help with schoolwork or other academic related stuff in the morning at our premises. The footballers are currently 35 in numbers including Allan mboto, Godwin, Desire, Yaya, Mandla mabena, Mava mto, Banzi mlhali, milton, jimmy, Tevor, Onele, Tawanda, Joseph, kalume patience, Guvamonbe blessing, Sly, Billy mkwashi, Mkize sandile,Ben chungu, Lambert, Sam chungu, Noku, Marc kahenga, Patie kalume...and a daily average of 3 newcomers members.  THIS SITE IS UNDER MAINTENANCE, WE APOLOGIZE FOR ANY INCONVENIENCE CAUSED
WHO WE ARE
Our purpose

booksnsports SA was officially launched as an NPO in South africa in 2011 as a collaborative effort among young members of a sport group in port elizabeth. The initiative was led by our founder Patient Kalume and co-founder Daniel Taokubong with the aim of diversifying the group's activities to create meaningful and fun activities for youths in the area including but not restricted to sport. With that a robust organization that included sport, art, academic follow up like tutoring and media was created. the founder, being a tutor and a fitness trainer decided to officially transform the mouvement into an NPO after noticing the need of an adequate education amongst the youth.

booksnsports SA strives to stand out with a different approach to development work and vouch for democracy and transparency. Experience has taught us that for a project to have long-term effects, it is crucial to have the beneficiaries partake in every aspect of the activities.

We come from Port elizabeth. 

Being that most of our members have grown up in Port elizabeth, we know first-hand the feeling of boredom and the feeling of having nothing meaningful to do. Idleness often is an enemy as threatening as poverty itself. Notwithstanding, port elizabeth is a vibrant, creative and lively area with strong and friendly communities. Rather the problem, especially for young people, is the lacking of tools, space and appropriate guidance to develop their creative interests. This reflects a general need for skill-developing and confidence-building, concerning education on other levels than merely the academic.

We engage the youth.

Although sport was the major foundation of our own personal development, we felt it was not by itself sufficient. The activities have therefore evolved into art,tutoring(youth that are still in schools) in area like mathematic and science , sport and an education program to pay school fees, provide mentoring sessions and individual academic follow up. These activities are good ways to sidestep destructive activities like criminality and substance abuse and also, it enhances the opportunities for youths to realize their potential and increase their chances and ability to build a great future for themselves.

It may be challenging but time has proven that it is possible.

booksnsports SA members are not exempted from the challenging situations that come from growing up in a slum. Over the years since it started as a soccer group and extending to where it is, a fully operating NPO, we have had the pleasure of witnessing some of our members growing from young, clueless but talented kids to self sufficient inspired adults. This image constantly reminds us to focus on participation and ownership as fundamentals of community-driven development. booksnsports is built upon the involvement of the youths, playing important roles both in everyday work and projects as well as in the overall managing of the organization.

Our Reach

Our organization continues to expand and develop. Our members meet at the greenfield ground for sports sessions every afternoon, six days a week and we have a daily attendance of about 30 youths . we gather as well with Members(about 20) that needs help with schoolwork or other academic related stuff in the morning at our premises. The footballers are currently 35 in numbers including Allan mboto, Godwin, Desire, Yaya, Mandla mabena, Mava mto, Banzi mlhali, milton, jimmy, Tevor, Onele, Tawanda, Joseph, kalume patience, Guvamonbe blessing, Sly, Billy mkwashi, Mkize sandile,Ben chungu, Lambert, Sam chungu, Noku, Marc kahenga, Patie kalume...and a daily average of 3 newcomers members.

THIS SITE IS UNDER MAINTENANCE, WE APOLOGIZE FOR ANY INCONVENIENCE CAUSED